+++ RayTracingDemo/GeometricObjects/Primitives/Plane.cpp
@@ -86,7 +86,7 @@ Plane::hit(const Ray& ray, double& tmin, ShadeRec& sr) const {
 }
 
 bool 
-Plane::shadow_hit(const Ray& ray, float& tmin) const
+Plane::shadow_hit(const Ray& ray, double& tmin) const
 {
 	// Specify whether to cast shadow or not
 	if (!shadows)
