+++ RayTracingDemo/GeometricObjects/Primitives/Plane.h
@@ -35,7 +35,7 @@ class Plane: public GeometricObject {
 		hit(const Ray& ray, double& tmin, ShadeRec& sr) const;
 
 		virtual bool 
-		shadow_hit(const Ray & ray, float & tmin) const;
+		shadow_hit(const Ray& ray, double& tmin) const;
 		
 	private:
