+++ RayTracingDemo/GeometricObjects/Primitives/Sphere.h
@@ -45,7 +45,7 @@ class Sphere: public GeometricObject {
 		hit(const Ray& ray, double& t, ShadeRec& s) const;	
 
 		virtual bool
-			shadow_hit(const Ray & ray, float & tmin) const;
+			shadow_hit(const Ray & ray, double& tmin) const;
 		
 	private:
